refactor: move 메서드 JspView로 이동
