+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/ModelAndView.java
@@ -1,6 +1,7 @@
 package webmvc.org.springframework.web.servlet;
 
-import java.util.Collections;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -23,11 +24,8 @@ public class ModelAndView {
         return model.get(attributeName);
     }
 
-    public Map<String, Object> getModel() {
-        return Collections.unmodifiableMap(model);
+    public void render(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        view.render(model, request, response);
     }
 
-    public View getView() {
-        return view;
-    }
 }
