+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/View.java
@@ -2,11 +2,10 @@ package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Map;
 
 public interface View {
+
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
 
-    String getPath();
 }
